+++ addon/options.js
@@ -260,6 +260,7 @@ class OptionsTabSelector extends React.Component {
           {option: Option, props: {type: "toggle", default: true, title: "Use BOM for CSV export", key: "useBomForCsvExport", tooltip: "Add UTF-8 BOM (Byte Order Mark) for Excel compatibility with non-Latin characters."}},
           {option: Option, props: {type: "toggle", title: "Enable SOQL Styling", key: "enableSoqlStyling", default: false, tooltip: "Enable syntax highlighting for SOQL keywords in the query input"}},
           {option: Option, props: {type: "toggle", title: "Enable SOQL Comments", key: "enableSoqlComments", default: false, tooltip: "Enable management of comments (-- and /* */) in the query input"}},
+          {option: Option, props: {type: "toggle", title: "Enable field/object autocomplete popup", key: "enableDataExportAutocomplete", default: false, tooltip: "Show the autocomplete popup for field and object suggestions when typing in the query input"}},
         ]
       },
       {